server<->order

+++ test.py
@@ -26,16 +26,17 @@ data_to_send = {
 }
 
 r = requests.post('http://localhost:3000/validateAndPrice',
-                 json=data_to_send)
+                  json=data_to_send)
 response_status = json.loads(r.text)["result"]["Status"]
 
 if int(response_status) == 1 or int(response_status) == 0:
     price_in_usd = json.loads(r.text)["result"]["Order"]["Amounts"]["Payment"]
 
-    get_bitpay_btc_usd_rate = request.urlopen(url="https://bitpay.com/api/rates/usd").read().decode("utf-8")
+    get_bitpay_btc_usd_rate = request.urlopen(
+        url="https://bitpay.com/api/rates/usd").read().decode("utf-8")
     usd_per_btc = json.loads(get_bitpay_btc_usd_rate)["rate"]
 
-    price = price_in_usd * 10**8 / usd_per_btc
+    price = int(price_in_usd * 10**8 / usd_per_btc)
 else:
     setattr(request, 'error_validate', 'error_validate')
     price = 0
